+++ src/components/Category/AuctionCategory/AuctionCategory.css.ts
