--- src/components/LeftSection/index.tsx
@@ -1,13 +1,0 @@
-import AuctionCategoryLeftSection from '@/components/LeftSection/components/AuctionCategoryLeftSection/AuctionCategoryLeftSection';
-import MobileAuctionCategoryLeftSection from '@/components/LeftSection/components/MobileAuctionCategoryLeftSection/MobileAuctionCategoryLeftSection';
-
-function LeftSection() {
-  return (
-    <>
-      <AuctionCategoryLeftSection />
-      <MobileAuctionCategoryLeftSection />
-    </>
-  );
-}
-
-export default LeftSection;
